% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{BarChartProp}
\alias{BarChartProp}
\title{Draw a stacked bar chart for cell-type proportions}
\usage{
BarChartProp(Prop, Color, TypeName = NULL, SampleName = NULL)
}
\arguments{
\item{Prop}{A matrix of the estimated cell type proportions to visualize in the bar chart}

\item{Color}{A character vector giving the color of all the cell types}

\item{TypeName}{A character vector giving the name of each cell type}

\item{SampleName}{A character vector giving the name of each sample}
}
\description{
The function \code{BarChartProp} draws the cell type proportions
}
\examples{

Cell_prop <- Prop(DIFseq_obj)
Celltype_color <- rainbow(ncol(Cell_prop))
BarChartProp(Cell_prop, Color = Celltype_color)

}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
