% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{DiffAbundance}
\alias{DiffAbundance}
\title{Conduct differential abundance analysis}
\usage{
DiffAbundance(
  DIFseqObj,
  ref = NULL,
  dim = c("Cond", "Batch", "Pair"),
  subset = NULL
)
}
\arguments{
\item{DIFseqObj}{A matrix, \code{\linkS4class{SingleCellExperiment}} or DIFseq object
containing feature-by-cell gene expression data.}

\item{ref}{An integer indicates the reference cell type. By default, the last cell type is regarded as the reference cell type}

\item{dim}{Dimension to be test. It has three possible options: \code{Batch}, \code{Cond} and `Pair`.}

\item{subset}{A numeric vector indicating the batches/conditions to be included in the comparison. By default,
consider all of the batches, conditions or pairs in differential abundance analysis}
}
\value{
A DIFseq object storing the overall and cell-type-specific test statistics
and p-values in the \code{estimation} slot.
}
\description{
The function \code{DiffAbundance} apply rigorous hypothesis testing to
identify differential abundance (DA) across batches or conditions
}
\details{
The `dim` argument specifies the group of cells to conduct the DA analysis.
In particular, `Batch` indicates to check DA between batches involved `subset` argument over all of the conditions;
`Cond` means to check DA between conditions specified by `subset` argument over all of the batches;
similarly, `Pair` means to check DA between different pairs involved in `subset` argument.
}
\examples{

DIFseq_obj <- DiffAbundance(DIFseq_obj, dim = "Cond")
DiffAbund(DIFseq_obj)

DIFseq_obj <- DiffAbundance(DIFseq_obj, dim = "Batch")
DiffAbund(DIFseq_obj)

}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
