% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{DiffExpression}
\alias{DiffExpression}
\title{Identify intrinsic genes and cell-type-specific DE genes}
\usage{
DiffExpression(DIFseqObj, fdr = 0.05, min.prop = 0.1)
}
\arguments{
\item{DIFseqObj}{A matrix, \code{\linkS4class{SingleCellExperiment}} or DIFseq object
containing feature-by-cell gene expression data.}

\item{fdr}{Bayesian false discovery rate of all intrinsic genes and cell-type-specific DE genes.}

\item{min.prop}{Minimum proportions of cells with non-zero read counts in a given group.
For intrinsic genes, it requires the proportions of non-zero counts in a cell type for a given gene to be greater than the threshold,}
}
\value{
A DIFseq object storing the identified intrinsic genes and cell-type-specific DE genes in the `estimation` slot
}
\description{
The function \code{DiffExpression} not only controls the Bayesian false discovery rate (FDR) to adjust for multiple hypothesis tests,
but also controls the minimum proportions of non-zero counts in a group of cells to ensure the biological meaning.
}
\details{
In differetianl expression analysis, there are `G * (K-1)` hypothesis tests to identify intrinsic genes, where `G` and `K` denote the number of genes and cell types.
Meanwhile, there are `G * (T-1) * K` hypothesis tests for cell-type-specific DE genes, where `T` represents the number of conditions.
Thus, we control the Bayesian FDR of all the `G * (T * K - 1)` tests to adjust for multiple hypotheses.
}
\examples{

DIFseq_obj <- DiffExpression(DIFseq_obj)
DE_genes <- DiffExp(DIFseq_obj)

}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
