% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIFseq-package.R
\name{DIFseq-package}
\alias{DIFseq-package}
\title{The DIFseq package}
\description{
The \pkg{DIFseq} package performs differential expression and differential
abundance analysis for the single-cell experiments collected under different conditions.
For details please see the vignettes \code{vignette("DIFseq_demo", package="DIFseq")} and
}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
