% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{DIFseq_MCEM}
\alias{DIFseq_MCEM}
\title{Implement MCEM algorithm for the posterior inference of the DIFseq model}
\usage{
DIFseq_MCEM(
  DIFseqObj,
  count = "counts",
  working_dir = getwd(),
  correction = TRUE
)
}
\arguments{
\item{DIFseqObj}{A matrix, \code{\linkS4class{SingleCellExperiment}} or DIFseq object
containing feature-by-cell gene expression data.}

\item{count}{A string indicating the name of count data matrix in DIFseq object}

\item{working_dir}{directory to store the parameter values at all of the iterations}

\item{correction}{A Boolean value indiacting whether to compute the corrected count data}
}
\value{
A \code{\linkS4class{DIFseq}} object with the parameter estimation.
}
\description{
The function \code{DIFseq_MCEM} implements an MCEM algorithm to fit the model of DIFferential inference for scRNA-seq data.
DIFseq is an interpretable Bayesian hierarchical model that closely follows the data-generating mechanism of
scRNA-seq experiments from multiple conditions. DIFseq can simultaneously correct batch effects, cluster cell types,
impute missing data caused by dropout events and detect differentially expressed genes without requiring a preliminary normalization step.
We develop an MCEM algorithm to conduct posterior inference for the BUSseq model. Here, we denote the batch number as B,
the gene number as G and the user-specific cell-type number as K.
}
\details{
This function implements our proposed MCEM algorithm to conduct
the posterior inference for the DIFseq model.
}
\examples{

DIFseq_obj <- DIFseq_MCEM(DIFseq_obj)

}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
