% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{HeatmapEffects}
\alias{HeatmapEffects}
\title{Draw heatmaps for the estimated effect size or count matrix}
\usage{
HeatmapEffects(matrix, ColSideColors, col = NULL, breaks = NULL)
}
\arguments{
\item{matrix}{A matrix of estimated parameters or a read count matrix to visualize on the heatmap.}

\item{col}{A vector of colors to fill each cell of the heatmap based on their corresponding values}

\item{breaks}{A set of breakpoints for the color key: must have one more breakpoints than the color key
and be in increasing order. Unsorted vectors will be sorted, with a warning.}

\item{color_bar}{A character vector of length ncol(x) containing the color names for a
horizontal side bar that may be used to annotate the columns of x}
}
\description{
The function \code{HeatmapEffects} enables to visualize the estimated parameter values or read count matrix
}
\examples{

Celltype_effects <- TypeEffects(DIFseq_obj)
Celltype_color <- rainbow(ncol(Celltype_effects))
HeatmapEffects(Celltype_effects, ColSideColors = Celltype_color)

}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
