% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\docType{methods}
\name{DIFseq-methods}
\alias{DIFseq-methods}
\alias{NumCond}
\alias{NumCond,DIFseq-method}
\alias{NumBatch}
\alias{NumBatch,DIFseq-method}
\alias{NumPair}
\alias{NumPair,DIFseq-method}
\alias{NumSample}
\alias{NumSample,DIFseq-method}
\alias{CondInd}
\alias{CondInd,DIFseq-method}
\alias{BatchInd}
\alias{BatchInd,DIFseq-method}
\alias{PairInd}
\alias{PairInd,DIFseq-method}
\alias{SampleInd}
\alias{SampleInd,DIFseq-method}
\alias{MiniBatch}
\alias{MiniBatch<-}
\alias{MiniBatch,DIFseq-method}
\alias{MiniBatch<-,DIFseq-method}
\alias{EarlyStop}
\alias{EarlyStop<-}
\alias{EarlyStop,DIFseq-method}
\alias{EarlyStop<-,DIFseq-method}
\alias{CheckLike}
\alias{CheckLike<-}
\alias{CheckLike,DIFseq-method}
\alias{CheckLike<-,DIFseq-method}
\alias{MCRep}
\alias{MCRep<-}
\alias{MCRep,DIFseq-method}
\alias{MCRep<-,DIFseq-method}
\alias{Ncores}
\alias{Ncores<-}
\alias{Ncores,DIFseq-method}
\alias{Ncores<-,DIFseq-method}
\alias{Seed}
\alias{Seed<-}
\alias{Seed,DIFseq-method}
\alias{Seed<-,DIFseq-method}
\alias{BIC_DIFseq}
\alias{BIC_DIFseq,DIFseq-method}
\alias{Prop}
\alias{Prop,DIFseq-method}
\alias{CellTypes}
\alias{CellTypes,DIFseq-method}
\alias{Baseline}
\alias{Baseline,DIFseq-method}
\alias{TypeEffects}
\alias{TypeEffects,DIFseq-method}
\alias{CondEffects}
\alias{CondEffects,DIFseq-method}
\alias{BatchEffects}
\alias{BatchEffects,DIFseq-method}
\alias{OverDisp}
\alias{OverDisp,DIFseq-method}
\alias{CellEffects}
\alias{CellEffects,DIFseq-method}
\alias{DropoutCoef}
\alias{DropoutCoef,DIFseq-method}
\alias{ImputedCounts}
\alias{ImputedCounts,DIFseq-method}
\alias{CorrectedCounts}
\alias{CorrectedCounts,DIFseq-method}
\alias{Intrinsic}
\alias{Intrinsic,DIFseq-method}
\alias{DiffExp}
\alias{DiffExp,DIFseq-method}
\alias{DiffAbund}
\alias{DiffAbund,DIFseq-method}
\alias{show}
\alias{show,DIFseq-method}
\title{Methods for extracting statistical inference results from a DIFseq object}
\description{
Get or set methods for Milo object slots. Generally speaking these methods
are used internally, but they allow the user to assign their own externally computed
values - should be used \emph{with caution}.
}
\section{Dimension information}{

In the following descriptions \code{x} is always a \linkS4class{DIFseq} object.
\describe{
\item{\code{NumCond(x)}:}{Return the number of conditions \code{T}.}
\item{\code{NumBatch(x)}:}{Return the number of batches \code{B}.}
\item{\code{NumPair(x)}:}{Return the number of batch-condition pairs \code{P}.}
\item{\code{NumSample(x)}:}{Return the number of samples \code{S}.}
\item{\code{BatchInd(x)}:}{Return the batch indices for all the cells.}
\item{\code{CondInd(x)}:}{Return the condition indices for all the cells.}
\item{\code{PairInd(x)}:}{Return the pair indices for all the cells.}
\item{\code{SampleInd(x)}:}{Return the sample indices for all the cells.}
 }
}

\section{Adjust the running parameters of MCEM algorithm}{

\describe{
\item{\code{MiniBatch(x)}:}{Return the mini-batch size of the MCEM algorithm.}
\item{\code{MiniBatch(x)<-value}:}{Set the mini-batch size of the MCEM algorithm.}
\item{\code{EarlyStop(x)}:}{Return the number of iterations allowing early stopping in the MCEM algorithm.}
\item{\code{EarlyStop(x)<-value}:}{Set the number of iterations allowing early stopping in the MCEM algorithm.}
\item{\code{CheckLike(x)}:}{Return the number of iterations checking the increase of logliklihood function in the MCEM algorithm.}
\item{\code{CheckLike(x)<-value}:}{Set the number of iterations checking the increase of logliklihood function in the MCEM algorithm.}
\item{\code{MCRep(x)}:}{Return the number of replicates in the Monte Carlo E step.}
\item{\code{MCRep(x)<-value}:}{Set the number of replicates in the Monte Carlo E step.}
\item{\code{Ncores(x)}:}{Return the number of cores for parallel computing in the MCEM algorithm.}
\item{\code{Ncores(x)<-value}:}{Set the number of cores for parallel computing in the MCEM algorithm.}
\item{\code{Seed(x)}:}{Return the seed of the MCEM algorithm.}
\item{\code{Seed(x)<-value}:}{Set the seed of the MCEM algorithm.}
}
}

\section{Estimated parameter values or results of posterior inference}{

\describe{
\item{\code{BIC_DIFseq(x)}:}{Return the Bayesian Informtion Criterion (BIC) value.}
\item{\code{Prop(x)}:}{Return a \code{S} by \code{K} matrix of the estimated sample-specific cell type proportions.}
\item{\code{CellTypes(x)}:}{Return a vector of length \code{N} indicating the estimated cell type labels for all cells.}
\item{\code{Baseline(x)}:}{Return a vector of length \code{G} giving the estimated baseline log-scale gene expression levels.}
\item{\code{TypeEffects(x)}:}{Return a \code{G} by \code{K} matrix of the estimated cell type effects compared with the first cell type.
Note that the first column is zero as the first cell type is taken as the reference cell type.}
\item{\code{CondEffects(x)}:}{Return a \code{G} by (\code{K * T}) matrix of the estimated condition effects compared with the first condition.
Note that the K columns are zero as the first condition is taken as the reference condition.}
\item{\code{BatchEffects(x)}:}{Return a \code{G} by \code{B} matrix of the estimated batch effects compared with the first batch.
Note that the first column is zero as the first batch is taken as the reference batch.}
\item{\code{OverDisp(x)}:}{Return a \code{G} by \code{B} matrix of the estimated over-dispersion.}
\item{\code{CellEffects(x)}:}{Return a vector of length \code{G} of the estimated cell size factors.}
\item{\code{DropoutCoef(x)}:}{Return a \code{B} by \code{2} matrix of the estimated dropout logistic coefficients.}
\item{\code{ImputedCounts(x)}:}{Return a \code{G} by \code{N} matrix of the imputed read counts.}
\item{\code{CorrectedCounts(x)}:}{Return a \code{G} by \code{N} matrix of the corrected read counts after removing batch effects and cell-size factor.}
}
}

\section{Differential inference results}{

\describe{
\item{\code{Intrinsic(x)}:}{Return a boolen vector of length \code{G} indicating the identified intrinsic genes across different cell types.}
\item{\code{DiffExp(x)}:}{Return a \code{G} by \code{K} matrix of the identified cell-type-specific DE genes across different conditions.}
\item{\code{DiffAbund(x)}:}{Return the statistics and log10 p-values in the differential abundance analysis.}
}


\describe{
\item{\code{show(x)}:}{Print information to the console regarding the \code{\linkS4class{DIFseq}} object.}
}
}

\examples{
show(DIFseq_obj)
}
\author{
Fangda Song, Kevin Y. Yip and Yingying Wei
}
